refactor parseOrigin
