+++ inlang/source-code/ide-extension/src/utilities/settings/getGitOrigin.ts
@@ -2,8 +2,7 @@
  * Gets the git origin url of the currently opened repository.
  */
 
-import { parseOrigin } from "@inlang/telemetry"
-import { findRepoRoot, _listRemotes } from "@lix-js/client"
+import { findRepoRoot, openRepository } from "@lix-js/client"
 import * as vscode from "vscode"
 import * as fs from "node:fs/promises"
 
@@ -11,18 +10,17 @@ export async function getGitOrigin() {
 	try {
 		const workspaceRoot = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath
 
-		// FIXME _listRemotes is deprecated. openRepo and then call repo.listRemotes instead!
-		const remotes = await _listRemotes({
-			fs,
-			dir: (
-				await findRepoRoot({
-					nodeishFs: fs,
-					path: workspaceRoot ?? process.cwd(),
-				})
-			)?.replace("file://", ""),
+		const repoRoot = await findRepoRoot({
+			nodeishFs: fs,
+			path: workspaceRoot ?? process.cwd(),
 		})
-		const gitOrigin = parseOrigin({ remotes })
-		return gitOrigin
+
+		if (!repoRoot) {
+			console.error("Failed to find repository root.")
+			return
+		}
+		const repo = await openRepository(repoRoot, { nodeishFs: fs })
+		return repo.getOrigin()
 	} catch (e) {
 		return undefined
 	}
