+++ inlang/source-code/paraglide/paraglide-js/src/services/telemetry/implementation.ts
@@ -3,10 +3,6 @@ import { PostHog } from "posthog-node"
 import type { TelemetryEvents } from "./events.js"
 
 // TODO add a project UUID to the tele.groups internal #196
-// import { getGitRemotes } from "../../utils/git.js"
-// import { parseOrigin } from "@inlang/telemetry"
-
-// export const gitOrigin = parseOrigin({ remotes: await getGitRemotes() })
 
 const posthog = new PostHog(publicEnv.PUBLIC_POSTHOG_TOKEN ?? "placeholder", {
 	host: "https://eu.posthog.com",
