+++ inlang/source-code/telemetry/src/index.ts
@@ -1,4 +1,2 @@
 export { telemetryBrowser } from "./implementation/browser.js"
 export { telemetryNode } from "./implementation/node.js"
-
-export { parseOrigin } from "./utilities/parseOrigin.js"
